+++ packages/billing_service/lib/billing_service.dart
@@ -4,3 +4,5 @@
 library;
 
 export 'src/billing_service_base.dart';
+export 'src/accounting_year_purchase.dart';
+export 'src/pawapay_checkout.dart';